# src/main/resources/swagger/api.yml
openapi: 3.0.3
info:
  title: Jasper KMS
  description: REST API for Jasper Knowledge Management Server
  version: '1.1.0'
  license:
    name: MIT License
    url: 'https://github.com/cjmalloy/jasper/blob/master/LICENSE'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/cjmalloy/Jasper/1.1.0
paths:
...
  /api/v1/plugin:
    get:
      tags:
        - Plugin
      operationId: getPlugin
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/plugin-qtag'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plugin'
        '304':
          description: Not Modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Plugin
      operationId: updatePlugin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plugin'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Plugin
      operationId: createPlugin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plugin'
        required: true
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - Plugin
      operationId: deletePlugin
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/plugin-qtag'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/plugin/list:
    get:
      tags:
        - Plugin
      operationId: getPluginList
      parameters:
        - name: tags
          in: query
          required: true
          schema:
            maxItems: 100
            minItems: 0
            type: array
            items:
              $ref: '#/components/schemas/plugin-qtag'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plugin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/plugin/page:
    get:
      tags:
        - Plugin
      operationId: getPluginPage
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 10
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
            default:
              - 'tag,ASC'
        - name: query
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/query'
        - name: modifiedAfter
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePlugin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/user:
    get:
      tags:
        - User
      operationId: getUser
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/user-qtag'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
        '304':
          description: Not Modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - User
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - User
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - User
      operationId: deleteUser
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/user-qtag'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  schemas:
    name:
      type: string
      description: Display name
      example: Name
      maxLength: 512
...
    origin:
      type: string
      pattern: '@[a-z]+(\.[a-z])*'
      default: ''
      description: Origin
      example: '@origin'
      maxLength: 64
...
    qtag:
      type: string
      pattern: '[_+]?[a-z]+(/[a-z]+)*(@[a-z]+(\.[a-z])*)?'
      description: Optionally Qualified Tag
      example: 'tag@origin'
      maxLength: 127
    query:
      type: string
      description: Tag Query
      pattern: '(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))|\(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))([ |]!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*)))*\))([ |:&](!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))|\(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))([ |]!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*)))*\)))*'
      example: '(politics|science):!funny'
      maxLength: 512
...
    plugin-tag:
      type: string
      pattern: '[_+]?plugin/[a-z]+(/[a-z]+)*'
      description: Plugin Tag
      example: 'plugin/thumbnail'
      maxLength: 64
    plugin-qtag:
      type: string
      pattern: '[_+]?plugin/[a-z]+(/[a-z]+)*(@[a-z]+(\.[a-z])*)?'
      description: Optionally Qualified Plugin Tag
      example: 'plugin/thumbnail@origin'
      maxLength: 127
    user-tag:
      type: string
      pattern: '[_+]user/[a-z]+(/[a-z]+)*'
      description: User Tag
      example: '+user/charlie'
      maxLength: 64
    user-qtag:
      type: string
      pattern: '[_+]user/[a-z]+(/[a-z]+)*(@[a-z]+(\.[a-z])*)?'
      description: Optionally Qualified User Tag
      example: '+user/charlie@origin'
      maxLength: 127
...
    Plugin:
      required:
        - tag
      type: object
      properties:
        tag:
          $ref: '#/components/schemas/plugin-tag'
        origin:
          $ref: '#/components/schemas/origin'
        name:
          $ref: '#/components/schemas/name'
        config:
          type: object
        defaults:
          type: object
        schema:
          type: object
        generateMetadata:
          type: boolean
        modified:
          type: string
          format: date-time
...
    User:
      required:
        - tag
      type: object
      properties:
        tag:
          $ref: '#/components/schemas/user-tag'
        origin:
          $ref: '#/components/schemas/origin'
        name:
          $ref: '#/components/schemas/name'
        readAccess:
          type: array
          items:
            $ref: '#/components/schemas/qtag'
        writeAccess:
          type: array
          items:
            $ref: '#/components/schemas/qtag'
        tagReadAccess:
          type: array
          items:
            $ref: '#/components/schemas/qtag'
        tagWriteAccess:
          type: array
          items:
            $ref: '#/components/schemas/qtag'
        modified:
          type: string
          format: date-time
...
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PagePlugin:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Plugin'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    NotFound:
      description: Not Found
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
...
    Conflict:
      description: Conflict
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    Forbidden:
      description: Forbidden
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'